Client packet rewrites done!
Former-commit-id: 2e07ea008ef060ef5bdaaaf51c3f3cd68e6ce180

+++ Intersect.Client/Classes/UI/Game/Hotbar/HotBar.cs
@@ -36,7 +36,7 @@ namespace Intersect.Client.UI.Game.Hotbar
             int x = 12;
             for (int i = 0; i < Options.MaxHotbar; i++)
             {
-                Items.Add(new HotbarItem(i, HotbarWindow));
+                Items.Add(new HotbarItem((byte)i, HotbarWindow));
                 Items[i].Pnl = new ImagePanel(HotbarWindow, "HotbarContainer" + i);
                 Items[i].Setup();
                 Items[i].KeyLabel = new Label(Items[i].Pnl, "HotbarLabel" + i);
